backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
sentence-transformers
transformers
torch
faiss-cpu
PyPDF2
requests
gunicorn